apps/api/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic>=2.11.7
pydantic-settings>=2.7.0
email-validator==2.2.0

# Supabase
supabase==2.28.0
python-jose[cryptography]==3.3.0

# Redis
redis==5.2.1

# HTTP
httpx==0.28.1

# Database (Phase 3+)
asyncpg==0.30.0

# Sentry (observability)
sentry-sdk[fastapi]==2.20.0

# Dev / test
pytest==8.3.4
pytest-asyncio==0.25.2
python-dotenv==1.0.1
